+++ R/pmml.ARIMA.R
@@ -1,7 +1,7 @@
 # PMML: Predictive Model Markup Language
 #
 # Copyright (c) 2009-2016, Zementis, Inc.
-# Copyright (c) 2016-2020, Software AG, Darmstadt, Germany and/or Software AG
+# Copyright (c) 2016-2021, Software AG, Darmstadt, Germany and/or Software AG
 # USA Inc., Reston, VA, USA, and/or its subsidiaries and/or its affiliates
 # and/or their licensors.
 #
@@ -83,6 +83,7 @@ pmml.ARIMA <- function(model,
                        app_name = "SoftwareAG PMML Generator",
                        description = "ARIMA Time Series Model",
                        copyright = NULL,
+                       model_version = NULL,
                        transforms = NULL,
                        missing_value_replacement = NULL,
                        ts_type = "statespace",
@@ -125,7 +126,7 @@ pmml.ARIMA <- function(model,
   pmml <- .pmmlRootNode()
 
   # PMML -> Header
-  pmml <- append.XMLNode(pmml, .pmmlHeader(description, copyright, app_name))
+  pmml <- append.XMLNode(pmml, .pmmlHeader(description, copyright, model_version, app_name))
 
   # PMML -> DataDictionary
   pmml <- append.XMLNode(pmml, .pmmlDataDictionary(field, transformed = transforms))