+++ R/pmml.itemsets.R
@@ -1,7 +1,7 @@
 # PMML: Predictive Model Markup Language
 #
 # Copyright (c) 2009-2016, Zementis, Inc.
-# Copyright (c) 2016-2020, Software AG, Darmstadt, Germany and/or Software AG
+# Copyright (c) 2016-2021, Software AG, Darmstadt, Germany and/or Software AG
 # USA Inc., Reston, VA, USA, and/or its subsidiaries and/or its affiliates
 # and/or their licensors.
 #
@@ -25,6 +25,7 @@ pmml.itemsets <- function(model,
                           app_name = "SoftwareAG PMML Generator",
                           description = "Frequent Itemsets Model",
                           copyright = NULL,
+                          model_version = NULL,
                           transforms = NULL, ...) {
   if (!inherits(model, "itemsets")) stop("Not a legitimate arules itemsets rules object")
 
@@ -35,7 +36,8 @@ pmml.itemsets <- function(model,
 
   # PMML -> Header
 
-  pmml <- append.XMLNode(pmml, .pmmlHeader(description, copyright, app_name))
+  pmml <- append.XMLNode(pmml, .pmmlHeader(description, copyright,
+                                           model_version, app_name))
 
   # PMML -> DataDictionary
   data.dictionary <- xmlNode("DataDictionary", attrs = c(numberOfFields = 2L))