+++ R/pmml.randomForest.R
@@ -1,7 +1,7 @@
 # PMML: Predictive Model Markup Language
 #
 # Copyright (c) 2009-2016, Zementis, Inc.
-# Copyright (c) 2016-2020, Software AG, Darmstadt, Germany and/or Software AG
+# Copyright (c) 2016-2021, Software AG, Darmstadt, Germany and/or Software AG
 # USA Inc., Reston, VA, USA, and/or its subsidiaries and/or its affiliates
 # and/or their licensors.
 #
@@ -58,6 +58,7 @@ pmml.randomForest <- function(model,
                               app_name = "SoftwareAG PMML Generator",
                               description = "Random Forest Tree Model",
                               copyright = NULL,
+                              model_version = NULL,
                               transforms = NULL,
                               missing_value_replacement = NULL,
                               parent_invalid_value_treatment = "returnInvalid",
@@ -99,7 +100,8 @@ pmml.randomForest <- function(model,
 
   # PMML -> Header
 
-  pmml <- append.XMLNode(pmml, .pmmlHeader(description, copyright, app_name))
+  pmml <- append.XMLNode(pmml, .pmmlHeader(description, copyright, model_version,
+                                           app_name))
 
   # PMML -> DataDictionary