+++ R/pmml.rules.R
@@ -1,7 +1,7 @@
 # PMML: Predictive Model Markup Language
 #
 # Copyright (c) 2009-2016, Zementis, Inc.
-# Copyright (c) 2016-2020, Software AG, Darmstadt, Germany and/or Software AG
+# Copyright (c) 2016-2021, Software AG, Darmstadt, Germany and/or Software AG
 # USA Inc., Reston, VA, USA, and/or its subsidiaries and/or its affiliates
 # and/or their licensors.
 #
@@ -45,6 +45,7 @@ pmml.rules <- function(model,
                        app_name = "SoftwareAG PMML Generator",
                        description = "Association Rules Model",
                        copyright = NULL,
+                       model_version = NULL,
                        transforms = NULL, ...) {
   if (!inherits(model, "rules")) stop("Not a legitimate arules rules object")
 
@@ -56,7 +57,8 @@ pmml.rules <- function(model,
 
   # PMML -> Header
 
-  pmml <- append.XMLNode(pmml, .pmmlHeader(description, copyright, app_name))
+  pmml <- append.XMLNode(pmml, .pmmlHeader(description, copyright,
+                                           model_version, app_name))
 
   # PMML -> DataDictionary