+++ R/xform_z_score.R
@@ -1,7 +1,7 @@
 # PMML: Predictive Model Markup Language
 #
 # Copyright (c) 2009-2016, Zementis, Inc.
-# Copyright (c) 2016-2020, Software AG, Darmstadt, Germany and/or Software AG
+# Copyright (c) 2016-2021, Software AG, Darmstadt, Germany and/or Software AG
 # USA Inc., Reston, VA, USA, and/or its subsidiaries and/or its affiliates
 # and/or their licensors.
 #
@@ -67,7 +67,7 @@
 #'
 #' @examples
 #'
-#' # Load the standard iris dataset, already built into R
+#' # Load the standard iris dataset
 #' data(iris)
 #'
 #' # First wrap the data